requirements.txt
google-generativeai
fastapi
pydantic
uvicorn[standard]